fix(models): load backbones in fp32 to stop the deberta-v3 fp16 NaN divergence
deberta-v3-base ships a float16 checkpoint and transformers>=5 loads in the checkpoint dtype, so fine-tuning it in raw fp16 (no loss scaling) diverged to NaN within an epoch and produced a degenerate surrogate. Force dtype=torch.float32 at every load site (training, CKA representations, the attack/transfer wrapper, the persisted-checkpoint dataset), catch a non-finite loss and retry once on CPU, and drop surrogates with non-finite similarity features before the regression.

+++ src/transfer_risk/datasets/surrogate_model_dataset.py
@@ -71,7 +71,11 @@ class SurrogateModelDataset(AbstractDataset[dict[str, Any], dict[str, Any]]):
         source = str(self._path)
         return {
             "kind": "transformer",
-            "model": AutoModelForSequenceClassification.from_pretrained(source),
+            # fp32 to match the training/inference path (fp16 weights are slow/unsupported on
+            # the CPU the attack sweep uses); see transfer_risk.modeling.
+            "model": AutoModelForSequenceClassification.from_pretrained(
+                source, dtype=torch.float32
+            ),
             "tokenizer": AutoTokenizer.from_pretrained(source),
         }
